convert getCommonTransactionBodyBytes and legacyReceiptStorageSecs to return long and remove casts

+++ hapi-fees/src/main/java/com/hedera/services/usage/crypto/CryptoOpsUsage.java
@@ -109,7 +109,7 @@ public class CryptoOpsUsage {
 	public FeeData cryptoCreateUsage(TransactionBody cryptoCreation, SigUsage sigUsage) {
 		var op = cryptoCreation.getCryptoCreateAccount();
 
-		int variableBytes = 0;
+		long variableBytes = 0;
 		variableBytes += op.getMemoBytes().size();
 		variableBytes += keySizeIfPresent(op, CryptoCreateTransactionBody::hasKey, CryptoCreateTransactionBody::getKey);
 		if (op.hasProxyAccountID()) {
@@ -121,7 +121,7 @@ public class CryptoOpsUsage {
 		var estimate = txnEstimateFactory.get(sigUsage, cryptoCreation, ESTIMATOR_UTILS);
 		/* Variable bytes plus two additional longs for balance and auto-renew period;
 		   plus a boolean for receiver sig required. */
-		estimate.addBpt((long) variableBytes + 2 * LONG_SIZE + BOOL_SIZE);
+		estimate.addBpt(variableBytes + 2 * LONG_SIZE + BOOL_SIZE);
 		estimate.addRbs((CRYPTO_ENTITY_SIZES.fixedBytesInAccountRepr() + variableBytes) * lifetime);
 		estimate.addNetworkRbs(BASIC_ENTITY_ID_SIZE * USAGE_PROPERTIES.legacyReceiptStorageSecs());
